Move out address codec implementations into submodules
