+++ tox.ini
@@ -2,8 +2,6 @@
 envlist = py27, py34, py35, py36, py37, pypy3
 
 [testenv]
-setenv =
-    PYTHONPATH = {toxinidir}:{toxinidir}/multiaddr
 commands = py.test --cov=./
 
 ; If you want to make tox run the tests with the same versions, create a
